pyproject.toml
[project]
name = "lms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "bcrypt>=4.3.0",
    "panda>=0.3.1",
    "pymongo>=4.14.0",
    "streamlit>=1.48.1",
]
